refactor(daemon): decouple memory classification
