+++ platform/daemon/src/memory-classification.ts
@@ -1,0 +1,76 @@
+import type { Database } from "bun:sqlite";
+
+const TYPE_HINTS: ReadonlyArray<readonly [string, string]> = [
+	["prefer", "preference"],
+	["likes", "preference"],
+	["want", "preference"],
+	["decided", "decision"],
+	["agreed", "decision"],
+	["will use", "decision"],
+	["learned", "learning"],
+	["discovered", "learning"],
+	["til ", "learning"],
+	["bug", "issue"],
+	["issue", "issue"],
+	["broken", "issue"],
+	["never", "rule"],
+	["always", "rule"],
+	["must", "rule"],
+] as const;
+
+export function inferType(content: string): string {
+	const lower = content.toLowerCase();
+	for (const [hint, type] of TYPE_HINTS) {
+		if (lower.includes(hint)) return type;
+	}
+	return "fact";
+}
+
+/** Decay-weighted score: pinned items always score 1.0 */
+export function effectiveScore(importance: number, createdAt: string, pinned: boolean): number {
+	if (pinned) return 1.0;
+	const ageDays = (Date.now() - new Date(createdAt).getTime()) / (1000 * 60 * 60 * 24);
+	return importance * 0.95 ** ageDays;
+}
+
+/** Check if content overlaps 70%+ with existing memories via FTS */
+export function isDuplicate(db: Database, content: string, agentId: string): boolean {
+	const words = content
+		.toLowerCase()
+		.split(/\W+/)
+		.filter((w) => w.length >= 3);
+	if (words.length === 0) return false;
+
+	try {
+		const ftsQuery = words.slice(0, 10).join(" OR ");
+		const rows = db
+			.prepare(
+				`SELECT m.content
+				 FROM memories_fts
+				 JOIN memories m ON memories_fts.rowid = m.rowid
+				 WHERE memories_fts MATCH ?
+				   AND m.is_deleted = 0
+				   AND m.agent_id = ?
+				 LIMIT 10`,
+			)
+			.all(ftsQuery, agentId) as Array<{ content: string }>;
+
+		const inputWords = new Set(words);
+		for (const row of rows) {
+			const rowWords = new Set(
+				row.content
+					.toLowerCase()
+					.split(/\W+/)
+					.filter((w) => w.length >= 3),
+			);
+			let overlap = 0;
+			for (const w of inputWords) {
+				if (rowWords.has(w)) overlap++;
+			}
+			if (overlap / inputWords.size >= 0.7) return true;
+		}
+	} catch {
+		// FTS table might not exist yet.
+	}
+	return false;
+}
