+++ platform/daemon/src/pipeline/summary-worker.ts
@@ -20,9 +20,9 @@ import { resolveDefaultBasePath } from "@signet/core";
 import { normalizeAndHashContent } from "../content-normalization";
 import type { DbAccessor, WriteDb } from "../db-accessor";
 import { countChanges } from "../db-helpers";
-import { inferType, isDuplicate } from "../hooks";
 import { getInferenceProvider } from "../llm";
 import { logger } from "../logger";
+import { inferType, isDuplicate } from "../memory-classification";
 import { loadMemoryConfig } from "../memory-config";
 import { IMMUTABLE_ARTIFACT_ERROR_PREFIX, writeSummaryArtifact } from "../memory-lineage";
 import { isNoiseSession } from "../session-noise";
